fix: rename planMode.ts and eliminate redundant workspace fetch
