+++ site/src/pages/AgentsPage/AgentCreatePage.tsx
@@ -18,7 +18,7 @@ import { AgentPageHeader } from "./components/AgentPageHeader";
 import { ChimeButton } from "./components/ChimeButton";
 import { WebPushButton } from "./components/WebPushButton";
 import { getModelOptionsFromConfigs } from "./utils/modelOptions";
-import { buildAgentChatPath } from "./utils/planMode";
+import { buildAgentChatPath } from "./utils/navigation";
 
 const lastModelConfigIDStorageKey = "agents.last-model-config-id";
 const nilUUID = "00000000-0000-0000-0000-000000000000";
