+++ site/src/pages/AgentsPage/utils/navigation.ts
